requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
charset-normalizer==2.0.12
Django==3.2.16
django-bootstrap3==21.1
django-paypal==2.0
frozenlist==1.2.0
future==0.18.2
idna==3.4
idna-ssl==1.1.0
importlib-metadata==2.1.3
liffpy==0.0.1a3
line-bot-sdk==2.2.1
multidict==5.2.0
Pillow==8.4.0
pytz==2022.6
requests==2.27.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.13
wincertstore==0.2
yarl==1.7.2
zipp==3.6.0
